tests/requirements.txt
pytest==8.1.1
pytest-depends==1.0.1
pytest-order==1.2.1
jinja2==3.1.3
pyopenssl==24.1.0
niquests==3.5.5
grpcio==1.62.1
opentelemetry-proto==1.24.0
